Updated course deletion logic

+++ SystemSchool/Datas/ConnectionDb.cs
@@ -189,13 +189,16 @@ namespace Datas
 
         public async Task DeleteRecordsInTableByIdAsync<T>(IEnumerable<int> recordIds) where T : class
         {
-            using (SqlConnection connection = new SqlConnection(ConfigurationManager.ConnectionStrings["sql"].ConnectionString)) 
+            if (recordIds.Any() && recordIds.Count() > 0)
             {
-                await connection.OpenAsync();
-                string idsToDelete = string.Join(",", recordIds);
-                string tableName = typeof(T).Name + "s";
-                string sqlQuery = $"DELETE FROM {tableName} WHERE Id IN ({idsToDelete})";
-                await connection.ExecuteAsync(sqlQuery);
+                using (SqlConnection connection = new SqlConnection(ConfigurationManager.ConnectionStrings["sql"].ConnectionString))
+                {
+                    await connection.OpenAsync();
+                    string idsToDelete = string.Join(",", recordIds);
+                    string tableName = typeof(T).Name + "s";
+                    string sqlQuery = $"DELETE FROM {tableName} WHERE Id IN ({idsToDelete})";
+                    await connection.ExecuteAsync(sqlQuery);
+                }
             }
         }
